- initial CsvTypeSourceTest
